Removed plone.allowdiscussion behavior from FAQ and FAQ Item.
This behavior is no longer available by default in Plone 6.1.
I also don't think it makes much sense for these two types, and allow_discussion was already set to False.
No upgrade step: this is only for new sites.  It is easy to manually turn on if a Site Administrator wants it.