Add icon expressions to our content type actions.
The Edit and View actions were not visible on the toolbar.

Create upgrade step and profile to for this.
We only want to add an icon_expr, but the import code is not smart enough to do only that.
So we need the other attributes as well.  This may override some settings, but I guess for this add-on it is okay.