Register indexer for getRawRelatedItems. Add catalog metadata.
I thought I would need this, so we can check the FAQ Item brains for related items.
But we get the object already, because we need some other info that is not available on the brain.
So this indexer and metadata are not needed here.
Can still be nice as example for a possible Plone improvement.